requirements.txt
arrow==1.1.1
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
tenacity==8.2.3
tmdbv3api==1.7.5
tqdm==4.65.0
trakt.py==4.4.0
typing-extensions>=4.6.0
urllib3==1.26.6
